server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
alembic==1.13.2
psycopg[binary]==3.2.13
pytest==8.3.2
httpx==0.27.2
python-dotenv==1.0.0
SQLAlchemy==2.0.46
